requirements.txt
PyExecJS==1.5.1
requests==2.23.0
Pillow==7.1.2
